zlib1g-dev dependency needed on Debian
The package zlib1g-dev needed to be added to the package install line for Debian/Ubuntu. I tried building the project and installed the required packages above; however, when I did so the ./configure failed to run due to a missing zlib requirement. Adding the above package resolved the issue.